Add - non-functional - openlilylib.svg icon
The preference dialog doesn't display it but instead complains
about links not found (although I can see in the file that they are
defined.
I don't know much about SVG and Inkscape, so I can't really judge this.
Of course the issue has to be resolved before this stuff can be merged.